backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai==0.2.2
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
gunicorn==21.2.0